changed db class for study_id in study and event tables to be REAL instead of INTEGER to match changes to code in r scripts, added step to populate the study_id col in event table with the value in the study's study table

+++ db/clean_study.r
@@ -49,7 +49,7 @@ if(interactive()) {
   .seed <- ag$seed
   .test <- as.logical(ag$test)
   rd <- is_rstudio_project$make_fix_file(.script)
-  .studyid <- as.integer(ag$studyid)
+  .studyid <- as.numeric(ag$studyid)
     
   source(rd('funs/input_parse.r'))
